#133 Separated utills definition from declaration
